requirements.txt
python3
python-docx=0.8.10
